chore(project): add update_snapshot_versions.js script
